% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_string.R
\name{is_string}
\alias{is_string}
\title{is_string}
\usage{
is_string(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Test if object is a string
}
\details{
Tests if \code{x} is a string, i.e. a lenght-1 charactaer vector.
}
