% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_detect_unbalanced.R
\name{str_detect_unbalanced}
\alias{str_detect_unbalanced}
\title{Detect unbalanced (quotes)}
\usage{
str_detect_unbalanced(string, quotes = c("'", "\\""))
}
\arguments{
\item{string}{character; string}

\item{quotes}{character; characters used for quiting}
}
\value{
logical
}
\description{
Detects whether strings contain unbalanced quites
}
\details{
Determines whether there are unb
Unbalanced quotes are determined by an odd number of quotes once all escaped
quotes are removed.
}
\examples{
str_detect_unbalanced( character(0) ) # locical(0)
  str_detect_unbalanced( "" )           # FALSE
  str_detect_unbalanced( "'" )          # TRUE

  "foo \\\\' bar" \%>\% cat
}
\references{
Stackoverflow
}

