% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_escape.R
\name{str_escape}
\alias{str_escape}
\alias{str_escape_dot}
\title{Escape dot(.) in a string}
\usage{
str_escape(string, x, escape = "\\\\")

str_escape_dot(string, x = ".", escape = "\\\\")
}
\arguments{
\item{string}{string}

\item{x}{string to be escaped, usually a single character.}
}
\description{
Replaces \code{.} with \code{\\.} in strings.
}
\details{
\code{str_escape} escapes \strong{all} instances of \code{x} in a string by \code{\\}. This
is handy when working with regular expressions, file descriptors, etc. where
certain characters have special meaning. By escaping, the character will
match litereally instead of using their special meanings.

\code{str_escape_dot} wraps \code{str_escape} using \code{x='.'} and \code{escape='\\'}
}
\examples{
  str_escape( "path/to/file", "/" )
  str_escape( "path/to/file", "/", '\\\\' )

  str_escape_dot("hello.world")   # "hello\\\\.world"
  str_escape_dot("foo.bar.baz")   # "foo\\\\.bar\\\\.baz"

}
