% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_pad.R
\name{str_rpad}
\alias{str_rpad}
\alias{str_lpad}
\alias{str_pad_max}
\alias{str_rpad_max}
\alias{str_lpad_max}
\title{pad strings on one right or left}
\usage{
str_rpad(string, width, pad = " ")

str_lpad(string, width, pad = " ")

str_pad_max(string, side = c("left", "right", "both"), pad = " ")

str_rpad_max(string, pad = " ")

str_lpad_max(string, pad = " ")
}
\arguments{
\item{string}{string to pad}

\item{width}{integer; minimum number of characters for the final string}

\item{pad}{single character string}

\item{side}{character; one of `left`, `right` or `both`}
}
\description{
pad strings on one right or left
}
\details{
These are convenience wrappers around [stringr::str_pad()]
}
