% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_snip.R
\name{str_snip}
\alias{str_snip}
\title{Snip portion from string}
\usage{
str_snip(string, pattern)
}
\arguments{
\item{string}{string}

\item{pattern}{pattern to remove from the string}
}
\description{
Removes a pattern from a string
}
\details{
`str_replace( string, pattern, '' )
}
