% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_split_ws_nonquote.R
\name{str_split_ws_nonquote}
\alias{str_split_ws_nonquote}
\title{Split on white-space removing quotes}
\usage{
str_split_ws_nonquote(x, quotes = c("'", "\\""))
}
\arguments{
\item{x}{character or string; (required)}

\item{quotes}{character or string; character interpreted as quotes. The
default is: ' *single-quote and " *double-quote}
}
\value{
character vector
}
\description{
Splits a string on whitespace but ignoring whitespace inside quotes
}
\details{
\code{str_split_ws_nonquote} splits strings into a vector elements delineated
by whitespace and ignoring whitespace enclosed in quotes.  It also removes
those quotes from strings.

\code{str_split_ws_nonquote} oses not (yet?) allow for different opening and
closing characters.
}
\examples{
character(0) \%>\% str_split_ws_nonquote()  # character(0)
  "" \%>\% str_split_ws_nonquote()            # "" - no splitting
  "foo" \%>\% str_split_ws_nonquote()         # "foo"
  "foo bar" \%>\% str_split_ws_nonquote       # "foo" "bar"
  "'foo' bar" \%>\% str_split_ws_nonquote     # "foo" "bar"
  "'foo' \\"bar\\"" \%>\% str_split_ws_nonquote # "foo" "bar"
  "'foo bar" \%>\% str_split_ws_nonquote      #
}
\author{
Decision Patterns from the base.tools packages; used with permission.
}

