% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_escape_dot.R
\name{str_escape_dot}
\alias{str_escape_dot}
\title{Escape dot(.) in a string}
\usage{
str_escape_dot(string)
}
\arguments{
\item{string}{string}
}
\description{
Replaces \code{.} with \code{\\.} in strings.
}
\details{
This escpaes \strong{all} instances in single-dot in strings. This is handy when
working with regular expressions where a period/dot is used to mean any
character. By escaping, the regular expression will match on a literal
period.

\code{str_replace_dot} is simply:\preformatted{ `str_replace_all(string, fixed("."), "\\.")``
}
}
\examples{
  str_escape_dot("hello.world")   # "hello\\\\.world"
  str_escape_dot("foo.bar.baz")   # "foo\\\\.bar\\\\.baz"

}
