% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_prefix.R
\name{str_prefix}
\alias{str_prefix}
\alias{str_suffix}
\title{Add prefix or suffix to a string}
\usage{
str_prefix(x, prefix, sep = ".")

str_suffix(x, suffix, sep = ".")
}
\arguments{
\item{x}{character vector or object with a names attribute}

\item{prefix}{string;}

\item{sep}{character vector to separate \code{prefix} from \code{x}. The
default is "."}

\item{suffix}{character;}
}
\value{
character vector with each element of \code{x} prefixed or sufffixed as
specified.
}
\description{
Adds a prefix or suffix to strings
}
\details{
This is a convenient wrapper for paste( prefix, x, sep ) and
paste( x, suffix, sep).
}
\examples{
  lets <- letters[1:5]
  str_prefix( lets, "letter" )
  str_prefix( lets, "letter", sep="_" )

  str_suffix( lets, "letter" )
  str_suffix( lets, "letter", sep="_" )

}

