% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_prefix.R
\name{str_prefix}
\alias{str_prefix}
\alias{str_suffix}
\title{Add prefix or suffix to a string}
\usage{
str_prefix(x, prefix, sep = getOption("str_prefix_separator",
  getOption("str_seperator", ".")))

str_suffix(x, suffix, sep = getOption("str_prefix_separator",
  getOption("str_seperator", ".")))
}
\arguments{
\item{x}{character vector or object with a names attribute}

\item{prefix}{string;}

\item{sep}{character vector to separate \code{prefix} from \code{x}. The
default is "." See Details.}

\item{suffix}{character;}
}
\value{
character vector with each element of \code{x} prefixed or sufffixed as
specified.
}
\description{
Adds a prefix or suffix to strings
}
\details{
This is a convenient wrapper for \code{paste( x, prefix, sep )} and
\code{paste( x, suffix, sep)}.

The separator is looked in the options list for: \code{str_prefix_separator} /
\code{str_suffux_separator} followed by \code{str_separator}. If those are not found
\code{.} is used.
}
\examples{

  lets <- letters[1:5]
  str_prefix( lets, "letter" )
  str_prefix( lets, "letter", sep="_" )

  str_suffix( lets, "letter" )
  str_suffix( lets, "letter", sep="_" )

}
